+++ lib/src/core/core.dart
@@ -1,3 +1,4 @@
+export 'package:geodesy/src/math/to_radian.dart';
 export 'package:geodesy/src/core/bearing_between_two_geo_points.dart';
 export 'package:geodesy/src/core/bounding_box.dart';
 export 'package:geodesy/src/core/points_in_range.dart';
@@ -5,6 +6,7 @@ export 'package:geodesy/src/core/get_rectangle_bounds.dart';
 export 'package:geodesy/src/core/get_polygon_intersection.dart';
 export 'package:geodesy/src/core/find_polygon_centroid.dart';
 export 'package:geodesy/src/core/geo_points.dart';
+export 'package:geodesy/src/core/vincenty_distance_calculation.dart';
 export 'package:geodesy/src/core/destination_point_by_distance_and_bearing.dart';
 export 'package:geodesy/src/core/cross_track_distance_to.dart';
 export 'package:latlong2/latlong.dart' hide pi;